Split list item component off from SnippetsController
